Refactor underlying pragma methods to parent directory
