+++ pyteal/pragma/__init__.py
@@ -1,0 +1,6 @@
+from pyteal.pragma.pragma import is_valid_compiler_version, pragma
+
+__all__ = [
+    "is_valid_compiler_version",
+    "pragma",
+]
